Fix broken 64bit test: int->size_t

+++ test/compilable/interpret3.d
@@ -1077,7 +1077,7 @@ void baqop(int n, ref int[] fongo) {
         fongo[0]++;
     }
 }
-int bug5258c() {
+size_t bug5258c() {
     Foo5258c qq;
     qq.r = new int[30];
     baqop(1, qq.r);
